+++ docs/release-notes.rst
@@ -35,6 +35,7 @@ Version 2.2.0, 26 Mar 2019
 * Define kernel-type values (`ACUMOS-2561 <https://jira.acumos.org/browse/ACUMOS-2561>`_)
 * Extend data and APIs for catalog features (`ACUMOS-2569 <https://jira.acumos.org/browse/ACUMOS-2569>`_)
 * Drop peer and solution groups (`ACUMOS-2590 <https://jira.acumos.org/browse/ACUMOS-2590>`_)
+* Query project, notebook and pipeline (`ACUMOS-2595 <https://jira.acumos.org/browse/ACUMOS-2595>`_)
 * Revise access type codes (`ACUMOS-2610 <https://jira.acumos.org/browse/ACUMOS-2610>`_)
 * Add Swagger library to display size constraints from JPA annotations in generated UI
 * Upgrade to Spring-Boot version 2.1.3.RELEASE